Add missing tests for auth and rename files
