+++ backend/events/tests/test_event_admin_forms.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 """
-Testing for proper event admin form cleaning
+Testing for proper event admin form cleaning.
 """
 
 import pytest
@@ -12,6 +12,7 @@ from events.models import Event
 
 pytestmark = pytest.mark.django_db
 
+
 class TestEventAdminForm(EventAdminForm):
     class Meta:
         model = Event
@@ -19,59 +20,58 @@ class TestEventAdminForm(EventAdminForm):
 
 
 # Ensure that when .clean() is called, an error is thrown
-# if there is invalid data. Do nothing otherwise. 
+# if there is invalid data. Do nothing otherwise.
+
 
 def test_good_phys_event() -> None:
     test_address = EventLocationFactory.create()
 
     form = TestEventAdminForm(EventAdminForm)
     form.cleaned_data = {
-        "location_type" : "physical",
-        "physical_location" : test_address,
-        "online_location_link" : ""
+        "location_type": "physical",
+        "physical_location": test_address,
+        "online_location_link": "",
     }
 
     assert form.clean()
 
-    
 
 def test_bad_phys_event() -> None:
     test_url = "https://example.com"
 
     form = TestEventAdminForm(EventAdminForm)
     form.cleaned_data = {
-        "location_type" : "physical",
-        "physical_location" : "", 
-        "online_location_link" : test_url
+        "location_type": "physical",
+        "physical_location": "",
+        "online_location_link": test_url,
     }
 
     with pytest.raises(ValidationError):
         form.clean()
 
-    
 
 def test_good_online_event() -> None:
     test_url = "https://example.com"
 
     form = TestEventAdminForm(EventAdminForm)
-    form.cleaned_data = {    
-        "location_type" : "online",
-        "physical_location" : "",
-        "online_location_link" : test_url
+    form.cleaned_data = {
+        "location_type": "online",
+        "physical_location": "",
+        "online_location_link": test_url,
     }
 
     assert form.clean()
-    
+
 
 def test_bad_online_event() -> None:
     test_address = EventLocationFactory.create()
 
     form = TestEventAdminForm(EventAdminForm)
     form.cleaned_data = {
-        "location_type" : "online", 
-        "physical_location" : test_address, 
-        "online_location_link" : ""
+        "location_type": "online",
+        "physical_location": test_address,
+        "online_location_link": "",
     }
-        
+
     with pytest.raises(ValidationError):
         form.clean()
